+++ CHANGES.md
@@ -15,10 +15,11 @@ Fast and flexible Python library for text tables.
 # Changes
 
 * 2.6.1 (2023-05-05)
-  * Simplified code.
-  * Corrected code typehints.
-  * Updated docs
+  * Fixed column container not updating width of columns correctly.
+  * Simplified header rendering code.
+  * Corrected typehints.
   * Added more tests.
+  * Updated docs.
 
 
 * 2.6.0 (2023-04-12)